+++ src/core/Wyam.Core/Execution/Pipeline.cs
@@ -7,18 +7,25 @@ namespace Wyam.Core.Execution
 {
     public class Pipeline : IPipeline
     {
-        public IModuleList ReadModules { get; } = new ModuleList();
+        /// <inheritdoc/>
+        public IModuleList InputModules { get; } = new ModuleList();
 
+        /// <inheritdoc/>
         public IModuleList ProcessModules { get; } = new ModuleList();
 
-        public IModuleList RenderModules { get; } = new ModuleList();
+        /// <inheritdoc/>
+        public IModuleList TransformModules { get; } = new ModuleList();
 
-        public IModuleList WriteModules { get; } = new ModuleList();
+        /// <inheritdoc/>
+        public IModuleList OutputModules { get; } = new ModuleList();
 
+        /// <inheritdoc/>
         public HashSet<string> Dependencies { get; } = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
 
+        /// <inheritdoc/>
         public bool Isolated { get; set; }
 
+        /// <inheritdoc/>
         public bool AlwaysProcess { get; set; }
     }
 }